requirements.txt
google-api-python-client==1.10.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
oauth2client==4.1.3
psycopg2-binary==2.8.5
google_auth_oauthlib==0.4.1
